[C] Refactoring to multi-file functions
