+++ .idea/encodings.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding" defaultCharsetForPropertiesFiles="UTF-8">
-    <file url="file://$PROJECT_DIR$/src/func/functions.cpp" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/functions/functions.cpp" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/include/fmt.h" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/include/function_definitions.h" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/include/macro.h" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/src/include/main.h" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/include/struct.h" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/src/main.cpp" charset="UTF-8" />
     <file url="PROJECT" charset="UTF-8" />
