+++ src/functions/LibraryBook_Copy.cpp
@@ -1,0 +1,10 @@
+//
+// Created by 地铁DXTL on 2024/1/4.
+//
+#include "../include/function_definitions.h"
+
+int LibraryBook_Copy(Book &bk, Book bk_sample)
+{
+    bk = std::move(bk_sample);
+    return 0;
+}
