+++ src/functions/LibraryBook_Update.cpp
@@ -1,0 +1,9 @@
+//
+// Created by 地铁DXTL on 2024/1/4.
+//
+
+#include "../include/function_definitions.h"
+
+int LibraryBook_Update(Book &bk) {
+    return 0;
+}
